First iteration of is_prime()
Implementing a primality test function for my own learning purposes. I
am giving myself a computer science education.

This basic implementation works. We are performance optimizing here in a
couple different ways:

1. if n is an even number we bail out right away
2. count up from the bottom of the range to sqrt(n)

This cuts time complexity to O(sqrt(n)) but it's actually better than
that because half the number space is early bailout before we start
looping. There may be other tests we can do to bail out early.

Going to continue to play with this and implement further optimizations and
concurrent processing of the number space.

Signed-off-by: Jen Farwell <jenfarwell21@gmail.com>