an open question
I'm not sure which is faster in the number space, moving up from 3 to
calculate modulos and test for factors, or to move DOWN from sqrt(n). I
would need to count instructions. It's also possible that it all evens
out if the numbers get big enough. I'm not sure! I should test and find
out.

Signed-off-by: Jen Farwell <jenfarwell21@gmail.com>